✨: Moving `RealIpKeyExtractor` struct outside the main
